+++ linear-array-extra/test/Data/Array/Mutable/Linear/UnboxedSpec.hs
@@ -30,7 +30,7 @@ import Data.Unrestricted.Linear (unur)
 import qualified Data.Unrestricted.Linear as Ur
 import qualified Data.Vector.Unboxed as U
 import Linear.Array.Extra.TestUtils
-import Linear.Witness.Token (linearly)
+import Linear.Token.Linearly (linearly)
 import qualified Prelude.Linear as PL
 import qualified Test.Falsify.Generator as F
 import Test.Falsify.Predicate ((.$))
