+++ linear-witness-test-utils/test/Data/Vector/Mutable/Linear/WitnessSpec.hs
@@ -26,8 +26,8 @@ import qualified Data.Unrestricted.Linear as Ur
 import qualified Data.Vector as V
 import qualified Data.Vector.Mutable.Linear as LV
 import qualified Data.Vector.Mutable.Linear.Witness as LV
-import Linear.Witness.Token (linearly)
-import Linear.Witness.Token.TestUtils
+import Linear.Token.Linearly (linearly)
+import Linear.Token.Linearly.TestUtils
 import qualified Prelude.Linear as PL
 import qualified Test.Falsify.Generator as F
 import Test.Falsify.Predicate ((.$))
