+++ linear-witness/README.md
@@ -5,5 +5,5 @@ This is just a tentative workaround until [Linear Constraints](https://github.co
 
 ## Contents
 
-- [`Linear.Witness.Token`](./src/Linear/Witness/Token.hs) provides `Linearly` constraints and related combinators.
+- [`Linear.Token.Linearly`](./src/Linear/Token/Linearly.hs) provides `Linearly` constraints and related combinators.
 - [`Data.**.Linear.Witness`](./src/Data) provides allocation primitives for containers provided in `linear-base`
